Move vector implementation
