+++ crates/core_simd/src/vector.rs
@@ -1,3 +1,6 @@
+#[macro_use]
+mod vector_impl;
+
 mod float;
 mod int;
 mod uint;
@@ -30,102 +33,4 @@ pub trait Vector: sealed::Sealed {
     /// Generates a SIMD vector with the same value in every lane.
     #[must_use]
     fn splat(val: Self::Scalar) -> Self;
-}
-
-macro_rules! impl_vector_for {
-    ($simd:ident {type Scalar = $scalar:ident;}) => {
-        impl_vector_for! { $simd<1> { type Scalar = $scalar; type BitMask = u8; } }
-        impl_vector_for! { $simd<2> { type Scalar = $scalar; type BitMask = u8; } }
-        impl_vector_for! { $simd<4> { type Scalar = $scalar; type BitMask = u8; } }
-        impl_vector_for! { $simd<8> { type Scalar = $scalar; type BitMask = u8; } }
-        impl_vector_for! { $simd<16> { type Scalar = $scalar; type BitMask = u16; } }
-        impl_vector_for! { $simd<32> { type Scalar = $scalar; type BitMask = u32; } }
-    };
-    ($simd:ident<$lanes:literal> {type Scalar = $scalar:ident; type BitMask = $bitmask:ident; }) => {
-        impl sealed::Sealed for $simd<$lanes> {}
-
-        impl Vector for $simd<$lanes> {
-            type Scalar = $scalar;
-            const LANES: usize = $lanes;
-
-            type BitMask = $bitmask;
-
-            #[inline]
-            fn splat(val: Self::Scalar) -> Self {
-                Self::splat(val)
-            }
-        }
-    };
-}
-
-impl_vector_for! {
-    SimdUsize {
-        type Scalar = usize;
-    }
-}
-
-impl_vector_for! {
-    SimdIsize {
-        type Scalar = isize;
-    }
-}
-
-impl_vector_for! {
-    SimdI8 {
-        type Scalar = i8;
-    }
-}
-
-impl_vector_for! {
-    SimdI16 {
-        type Scalar = i16;
-    }
-}
-
-impl_vector_for! {
-    SimdI32 {
-        type Scalar = i32;
-    }
-}
-
-impl_vector_for! {
-    SimdI64 {
-        type Scalar = i64;
-    }
-}
-
-impl_vector_for! {
-    SimdU8 {
-        type Scalar = u8;
-    }
-}
-
-impl_vector_for! {
-    SimdU16 {
-        type Scalar = u16;
-    }
-}
-
-impl_vector_for! {
-    SimdU32 {
-        type Scalar = u32;
-    }
-}
-
-impl_vector_for! {
-    SimdU64 {
-        type Scalar = u64;
-    }
-}
-
-impl_vector_for! {
-    SimdF32 {
-        type Scalar = f32;
-    }
-}
-
-impl_vector_for! {
-    SimdF64 {
-        type Scalar = f64;
-    }
 }
